AbstractCreateJobDispatchedMessagesTest is a functional test only
